requirements.txt
asgiref==3.5.0
Django==3.1.7
djangorestframework==3.13.1
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2==2.8.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
